remove redundant check for existing designator
Since we're already checking designationCategories on creating blueprints, and checking if we can build stuff on placing blueprints, there really is no need for a check on the designationCategory when we're creating a copy in BuildableInfo.

This simplifies designator creation, and resolves the need for compatibility with Architect Sense, as we're no longer reliant on the actual resolved designators.